Made AdyenWeChatPayInternal a local binary target as before
