+++ AdyenUIHost/Application/Configuration.swift
@@ -15,11 +15,11 @@ internal enum Configuration {
     
     static let appName = "Adyen Demo"
     
-    static let amount = Payment.Amount(value: 17408, currencyCode: "EUR")
+    static let amount = Payment.Amount(value: 17408, currencyCode: "VND")
     
     static let reference = "Test Order Reference - iOS UIHost"
     
-    static let countryCode = "NL"
+    static let countryCode = "VN"
 
     static let returnUrl = "ui-host://"
