+++ AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-arm64_armv7/AdyenWeChatPayInternal.framework/Modules/AdyenWeChatPayInternal.swiftmodule/armv7-apple-ios.swiftinterface
@@ -1,0 +1,6 @@
+// swift-interface-format-version: 1.0
+// swift-compiler-version: Apple Swift version 5.3 (swiftlang-1200.0.29.2 clang-1200.0.30.1)
+// swift-module-flags: -target armv7-apple-ios10.0 -enable-objc-interop -enable-library-evolution -swift-version 5 -enforce-exclusivity=checked -O -module-name AdyenWeChatPayInternal
+@_exported import AdyenWeChatPayInternal
+import Foundation
+import Swift
