+++ AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-arm64_armv7/AdyenWeChatPayInternal.framework/Modules/module.modulemap
@@ -4,3 +4,8 @@ framework module AdyenWeChatPayInternal {
   export *
   module * { export * }
 }
+
+module AdyenWeChatPayInternal.Swift {
+    header "AdyenWeChatPayInternal-Swift.h"
+    requires objc
+}
