--- AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-arm64_i386_x86_64-simulator/AdyenWeChatPayInternal.framework/Headers/AdyenWeChatPayInternal.h
@@ -1,21 +1,0 @@
-//
-//  AdyenWeChatPayInternal.h
-//  AdyenWeChatPayInternal
-//
-//  Created by Mohamed Eldoheiri on 10/22/20.
-//
-
-#import <Foundation/Foundation.h>
-#import <AdyenWeChatPayInternal/WXApiObject.h>
-#import <AdyenWeChatPayInternal/WXApi.h>
-#import <AdyenWeChatPayInternal/WechatAuthSDK.h>
-
-//! Project version number for AdyenWeChatPayInternal.
-FOUNDATION_EXPORT double AdyenWeChatPayInternalVersionNumber;
-
-//! Project version string for AdyenWeChatPayInternal.
-FOUNDATION_EXPORT const unsigned char AdyenWeChatPayInternalVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <AdyenWeChatPayInternal/PublicHeader.h>
-
-
