--- AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-arm64_i386_x86_64-simulator/AdyenWeChatPayInternal.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module AdyenWeChatPayInternal {
-  umbrella header "AdyenWeChatPayInternal.h"
-
-  export *
-  module * { export * }
-}
