--- AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-arm64_i386_x86_64-simulator/AdyenWeChatPayInternal.framework/_CodeSignature/CodeResources
@@ -1,177 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/AdyenWeChatPayInternal.h</key>
-		<data>
-		uqeKCLk2D2ajjJ83lGRACI2MxXw=
-		</data>
-		<key>Headers/WXApi.h</key>
-		<data>
-		ZuiG048tpHcwD307C/d9shcyZNo=
-		</data>
-		<key>Headers/WXApiObject.h</key>
-		<data>
-		tnYxt4QMA9pohDm3ALcUjzvzMdM=
-		</data>
-		<key>Headers/WechatAuthSDK.h</key>
-		<data>
-		6beFSxsQAVAU2YA49kPcCypShO0=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		1OZPn4hjHYs3+2T1aqRtk/dQB10=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		RsUaqCoMH4mBRWuFSZPQ/8cE4Y8=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/AdyenWeChatPayInternal.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			uqeKCLk2D2ajjJ83lGRACI2MxXw=
-			</data>
-			<key>hash2</key>
-			<data>
-			Hdk68fSAYo2X/YK2Ph8sPnV6qk7USBPm6E+qZM95/EY=
-			</data>
-		</dict>
-		<key>Headers/WXApi.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			ZuiG048tpHcwD307C/d9shcyZNo=
-			</data>
-			<key>hash2</key>
-			<data>
-			5W7WUJg2IV6Mt/71MBqt1+xBLRvk81F1By7QiwllLBk=
-			</data>
-		</dict>
-		<key>Headers/WXApiObject.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			tnYxt4QMA9pohDm3ALcUjzvzMdM=
-			</data>
-			<key>hash2</key>
-			<data>
-			cgxxEUknGQu2BTGFDmaCm/ihF9KKBT6LD0819Sz0HKA=
-			</data>
-		</dict>
-		<key>Headers/WechatAuthSDK.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			6beFSxsQAVAU2YA49kPcCypShO0=
-			</data>
-			<key>hash2</key>
-			<data>
-			eX6ctZW1kLlhw/hTNNWOJPj0BpgwHGwje4CjdGBKA+c=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			RsUaqCoMH4mBRWuFSZPQ/8cE4Y8=
-			</data>
-			<key>hash2</key>
-			<data>
-			S0iio6H7m25tM4f8QNldJBlsD2m5LAcoXOEmAml4xHA=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
