+++ AdyenWeChatPayInternal/AdyenWeChatPayInternal.xcframework/ios-i386_x86_64-simulator/AdyenWeChatPayInternal.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module AdyenWeChatPayInternal {
+  umbrella header "AdyenWeChatPayInternal.h"
+
+  export *
+  module * { export * }
+}
+
+module AdyenWeChatPayInternal.Swift {
+    header "AdyenWeChatPayInternal-Swift.h"
+    requires objc
+}
